+++ src/api/tests/feature_tests/steps/steps.py
@@ -139,32 +139,6 @@ def when_make_request(
     )
 
 
-@when(
-    'I make a "{http_method}" request with "{header_name}" headers to the id in the location response header to the endpoint prefix "{endpoint}"'
-)
-def when_make_request_with_id(
-    context: Context, http_method: str, header_name: str, endpoint: str
-):
-    endpoint = endpoint.replace("<id>", context.response.headers.get("Location"))
-    context.response = make_request(
-        base_url=context.base_url,
-        http_method=http_method,
-        endpoint=endpoint,
-        headers=context.headers[header_name],
-    )
-    context.postman_step.request = PostmanRequest(
-        url=Url(
-            raw=context.response.url,
-            host=[context.base_url.rstrip("/")],
-            path=[endpoint],
-        ),
-        method=http_method,
-        header=[
-            HeaderItem(key=k, value=v) for k, v in context.headers[header_name].items()
-        ],
-    )
-
-
 @then('I receive a status code "{status_code}" with body')
 def then_response(context: Context, status_code: str):
     expected_body = parse_table(table=context.table, context=context)
@@ -221,7 +195,7 @@ def then_response(context: Context, status_code: str, list_to_check: str, count:
 
 
 @then(
-    'I receive a status code "{status_code}" with a "{entity_type}" search body reponse that contains'
+    'I receive a status code "{status_code}" with a "{entity_type}" search body response that contains'
 )
 def then_response(context: Context, status_code: str, entity_type: str):
     expected_body = parse_table(table=context.table, context=context)
@@ -279,8 +253,10 @@ def then_response(context: Context):
     )
 
 
-@given('I note the response field "{jsonpath}" as "{alias}"')
-def note_response_field(context: Context, jsonpath: str, alias: str):
-    context.notes[alias] = extract_from_response_by_jsonpath(
-        response=context.response.json(), jsonpath=jsonpath
-    )
+for decorator in (given, when, then):
+
+    @decorator('I note the response field "{jsonpath}" as "{alias}"')
+    def note_response_field(context: Context, jsonpath: str, alias: str):
+        context.notes[alias] = extract_from_response_by_jsonpath(
+            response=context.response.json(), jsonpath=jsonpath
+        )